Use separate job for Control Plane clone and build (#87)
This change decouples building of the Control Plane image from running
tests, which allows flaky tests to be re-run using "Rerun workflow from
failed".

This should also fix the issue with SSH using the wrong deploy key,
since the deploy key used to pull the Control Plane repository is in a
separate job.